pyproject.toml
[project]
name = "loregen"
version = "0.1.0"
description = ""
authors = [
    {name = "ishish.io",email = "ts@ishish.io"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "gradio (>=5.25.2,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph-sdk (>=0.1.61,<0.2.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "langgraph (>=0.3.33,<0.4.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langchain-anthropic (>=0.3.15,<0.4.0)",
    "boto3 (>=1.38.34,<2.0.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "loregen-common (>=0.1.2,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
readme = "README.md"